Fixed typing errors around JSON parsed result
This uses the zod library (which I just learned about at the Google Next conference, great to put new learnings into action!)
Also: years of experience is no longer covered by search. It was implemented wrong, as revealed by the type check. It is not clear that it should be included in the search anyway. Can confirm with UX later.